Reading and playing files now works

+++ software/zynq/SoundgatesZynq/src/software_sound_components/wavefileplayer.h
@@ -2,7 +2,13 @@
  * wavefileplayer.h
  *
  *  Created on: 11.11.2013
- *      Author: soundgates
+ *      Author: gwue
+ *
+ *
+ *  The wavefileplayer basically justs loads data from a file into an array and returns the desired amount of entries.
+ *  It doesn't do any conversions yet (ie. the file must be encoded with 32Bit Unsigned Integer Mono Channel).
+ *  Also, it does not strip header files, if there is are any. Therefore, the file should be saved as raw data without a header.
+ *
  */
 
 #ifndef WAVEFILEPLAYER_H_
@@ -12,9 +18,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-typedef struct {
+typedef struct
+{
 	char* data;
-	int arraysize;
+	long int arraysize;
 	int offset;
 	int loop;
 } wavefileplayer;
@@ -27,12 +34,12 @@ wavefileplayer* wavefileplayer_create_from_path(char* path, int loop);
 /**
  * Returns samples from the wavefile.
  * If loop is 1, will start again from the beginning if the end is reached.
- * Otherwise fill with zero //TODO: Depending on the encoding zero would be the maximum negative amplitude! This should be changed
+ * Otherwise fill with zeros //TODO: Depending on the encoding (U32) zero would be the maximum negative amplitude! This should be changed
  *
- * targetSize must be the size of the char array to fill (not the sample size)
+ * targetSize must be the size of the char array to fill (not the number of samples)
  */
-void wavefileplayer_getSamples(wavefileplayer* wfp, int targetSize, char* target);
-
+void wavefileplayer_getSamples(wavefileplayer* wfp, int targetSize,
+		char* target);
 
 /**
  * Frees the space associated with a given wavefileplayer
